PKA-66 review DB integer sizes

+++ database/migrations/tenant/2022_11_29_101342_create_delivery_noteables.php
@@ -8,7 +8,7 @@ return new class () extends Migration {
     public function up()
     {
         Schema::create('delivery_noteables', function (Blueprint $table) {
-            $table->unsignedBigInteger('delivery_note_id');
+            $table->unsignedInteger('delivery_note_id');
             $table->foreign('delivery_note_id')->references('id')->on('delivery_notes');
             $table->morphs('delivery_noteable');
             $table->timestampsTz();
